backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.23.8
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
